pyproject.toml
[project]
name = "infinite-loop-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jorge A Jimenez", email = "me@jorgeajimenez.com" }
]
requires-python = ">=3.12"
dependencies = [
    "earthengine-api==0.1.391",
    "flask==3.0.2",
    "google-cloud-aiplatform==1.42.1",
    "google-cloud-firestore==2.15.0",
    "google-cloud-logging==3.9.0",
    "google-cloud-secret-manager==2.18.2",
    "google-cloud-storage==2.14.0",
    "google-cloud-texttospeech==2.16.2",
    "gunicorn==21.2.0",
    "mock==5.1.0",
    "pytest==8.0.0",
    "pytest-mock==3.12.0",
    "python-dotenv==1.0.1",
    "requests==2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
